Racoo 74

See merge request da253/library!13

+++ modules/libstp-sensor-ir/include/IRSensor.hpp
@@ -8,13 +8,12 @@
 #include <set>
 #endif
 #include "hal/Analog.hpp"
-
 namespace libstp::sensors::ir {
     class IRSensor : public hal::analog::AnalogSensor
     {
     public:
-        int whiteThreshold;
-        int blackThreshold;
+        float whiteThreshold;
+        float blackThreshold;
         float whiteMean = 0.0f;
         float blackMean = 0.0f;
         float whiteStdDev = 1.0f;
@@ -23,7 +22,7 @@ namespace libstp::sensors::ir {
 
         explicit IRSensor(const int& port, float calibrationFactor);
 
-        void setCalibration(int newBlackThreshold, int newWhiteThreshold);
+        void setCalibration(float newBlackThreshold, float newWhiteThreshold);
 
         bool isOnWhite() const;
